requirements.txt
requests
Flask
ollama
flask-cors
azure-ai-inference